Format dates in videos JSON
